Remove polymorphism
