+++ src/libraries/System.Net.Http.Json/ref/System.Net.Http.Json.netcoreapp.cs
@@ -10,7 +10,7 @@ using System.Text;
 
 namespace System.Net.Http.Json
 {
-    public abstract partial class JsonContent : System.Net.Http.HttpContent
+    public sealed partial class JsonContent : System.Net.Http.HttpContent
     {
         protected override void SerializeToStream(System.IO.Stream stream, System.Net.TransportContext? context, System.Threading.CancellationToken cancellationToken) { }
         protected override System.Threading.Tasks.Task SerializeToStreamAsync(System.IO.Stream stream, System.Net.TransportContext? context, System.Threading.CancellationToken cancellationToken) { throw null; }
