+++ src/libraries/System.Net.Http.Json/src/System.Net.Http.Json.csproj
@@ -24,8 +24,6 @@ System.Net.Http.Json.JsonContent</PackageDescription>
     <Compile Include="System\Net\Http\Json\HttpClientJsonExtensions.Patch.cs" />
     <Compile Include="System\Net\Http\Json\HttpContentJsonExtensions.cs" />
     <Compile Include="System\Net\Http\Json\JsonContent.cs" />
-    <Compile Include="System\Net\Http\Json\JsonContentOfT.cs" />
-    <Compile Include="System\Net\Http\Json\JsonContentUntyped.cs" />
     <Compile Include="System\Net\Http\Json\LengthLimitReadStream.cs" />
   </ItemGroup>
